união dos endpoint da classe UserControler
