+++ DEVinCar.Api/Controllers/AddressesController.cs
@@ -1,7 +1,7 @@
 ﻿using DEVinCar.Api.Models;
 using DEVinCar.Api.Data;
 using DEVinCar.Api.DTOs;
-using Microsoft.EntityFrameworkCore;
+
 using Microsoft.AspNetCore.Mvc;
 
 namespace DEVinCar.Api.Controllers;
